Add filtering pagination sorting and exception handling

+++ src/main/java/com/rajat/smartvehicletrackingsystem/service/VehicleServiceImpl.java
@@ -1,9 +1,13 @@
 package com.rajat.smartvehicletrackingsystem.service;
 
 import com.rajat.smartvehicletrackingsystem.entity.Vehicle;
+import com.rajat.smartvehicletrackingsystem.exception.VehicleNotFoundException;
 import com.rajat.smartvehicletrackingsystem.repository.VehicleRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -20,8 +24,8 @@ public class VehicleServiceImpl implements VehicleService{
 
     @Override
     public Vehicle updateVehicle(Long id, Vehicle vehicle) {
-        Vehicle existingVehicle = vehicleRepository.findById(id).orElse(null);
-        if (existingVehicle != null) {
+        Vehicle existingVehicle = vehicleRepository.findById(id).orElseThrow(() ->
+                new VehicleNotFoundException("Vehicle not found with id " + id));
 
             existingVehicle.setVehicleNumber(vehicle.getVehicleNumber());
             existingVehicle.setOwnerName(vehicle.getOwnerName());
@@ -33,32 +37,53 @@ public class VehicleServiceImpl implements VehicleService{
             existingVehicle.setStatus(vehicle.getStatus());
 
             return vehicleRepository.save(existingVehicle);
-        }
-        return null;
     }
     @Override
-    public List<Vehicle> getAllVehicles() {
-        return vehicleRepository.findAll();
+    public Page<Vehicle> getAllVehicles(int page, int size, String sortBy, String sortDir) {
+        Sort sort = sortDir.equalsIgnoreCase("desc")
+                ?Sort.by(sortBy).descending()
+                : Sort.by(sortBy).ascending();
+        Pageable pageable = PageRequest.of(page, size, sort);
+        return vehicleRepository.findAll(pageable);
     }
     @Override
     public Vehicle getVehicleById(Long id) {
-        return vehicleRepository.findById(id).orElse(null);
+        return vehicleRepository.findById(id).orElseThrow(() ->
+        new VehicleNotFoundException("vehicle not found with id " + id));
     }
     @Override
     public void deleteVehicleById(Long id) {
+        if (!vehicleRepository.existsById(id)) {
+            throw new VehicleNotFoundException("vehicle not found with id " + id);
+        }
         vehicleRepository.deleteById(id);
     }
     @Override
-    public List<Vehicle> getVehiclesByStatus(String status) {
-        return vehicleRepository.findByStatus(status);
+    public Page<Vehicle> getVehiclesByStatus(
+            String status,
+            int page,
+            int size,
+            String sortBy,
+            String sortDir
+    ){
+        Sort sort = sortDir.equalsIgnoreCase("desc")
+                ?Sort.by(sortBy).descending()
+                : Sort.by(sortBy).ascending();
+        Pageable pageable = PageRequest.of(page, size, sort);
+        return vehicleRepository.findByStatus(status, pageable);
     }
     @Override
-    public List<Vehicle> getVehiclesByOwnerName(String ownerName) {
-        return vehicleRepository.findAllByOwnerName(ownerName);
+    public Page<Vehicle> getVehiclesByOwnerName(String ownerName, int page, int size, String sortBy, String sortDir) {
+        Sort sort = sortDir.equalsIgnoreCase("desc")
+                ?Sort.by(sortBy).descending()
+                : Sort.by(sortBy).ascending();
+        Pageable pageable = PageRequest.of(page, size, sort);
+        return vehicleRepository.findAllByOwnerName(ownerName, pageable);
     }
     @Override
     public Vehicle getVehicleByVehicleNumber(String vehicleNumber) {
         return vehicleRepository.findByVehicleNumber(vehicleNumber)
-                .orElse(null);
+                .orElseThrow(() ->
+                        new VehicleNotFoundException("Vehicle not found with vehicle number " + vehicleNumber));
     }
 }
